Use globalSourceCache for contract upload routes
